requirements.txt
streamlit
pandas
numpy
plotly
paho-mqtt
fastapi
uvicorn
streamlit-autorefresh